Move methods behind interfaces
