+++ interfaces/projectupdater.ts
@@ -1,8 +1,7 @@
 import { TeamProject } from "azure-devops-node-api/interfaces/CoreInterfaces";
-import { IProject, IProjectPermission, IPermission } from "./configurationreader";
+import { IProject, IProjectPermission } from "./configurationreader";
 import { IGraphHelper } from "./graphhelper";
 import { IProjectHelper } from "./projecthelper";
-import { GraphGroup } from "azure-devops-node-api/interfaces/GraphInterfaces";
 
 export interface IProjectUpdater {
 
@@ -12,8 +11,6 @@ export interface IProjectUpdater {
     getProject(name: string): Promise<TeamProject>;
     createProject(project: IProject): Promise<TeamProject>;
     updateProject(project: IProject): Promise<TeamProject>;
-    updateGroupPermissions(projectName: string, group: GraphGroup, permissions: IPermission[]): Promise<void>;
-    updateGroupMembers(members: string[], group: GraphGroup): Promise<void>;
     updatePermissions(project: TeamProject, policy: IProjectPermission): Promise<void>;
 
 }
