+++ updaters/projectupdater.ts
@@ -82,66 +82,6 @@ export class ProjectUpdater implements IProjectUpdater {
 
     }
 
-    public async updateGroupPermissions(projectName: string, group: GraphGroup, permissions: IPermission[]): Promise<void> {
-
-        const debug = this.debugLogger.extend("updateGroupPermissions");
-
-        const groupProvider: IGroupProvider = await this.securityHelper.getGroupProvider("ms.vss-admin-web.org-admin-groups-permissions-pivot-data-provider", projectName, group);
-
-        for (const permission of permissions) {
-        
-            const targetPermission: ISubjectPermission = groupProvider.subjectPermissions.filter((i) => i.displayName === permission.name)[0];
-
-            if (!targetPermission) {
-
-                throw new Error(`Permission <${permission.name}> not found`);
-
-            }
-
-            // Some magic to address JSON enum parsing issue
-            // To be fixed with configuration reader refactoring
-            const type: PermissionType = PermissionType[permission.type.toString() as keyof typeof PermissionType];
-
-            // Skip updating identical explicit permission
-            if (targetPermission.explicitPermissionValue === type) {
-
-                debug(`Permission <${permission.name}> (${permission.type}) is identical`);
-
-                continue;
-            }
-
-            debug(`Configuring <${permission.name}> (${permission.type}) permission`);
-
-            const updatedPermission: any = await this.securityHelper.setGroupAccessControl(groupProvider.identityDescriptor, targetPermission, type);
-
-        }
-
-    }
-
-    public async updateGroupMembers(members: string[], group: GraphGroup): Promise<void> {
-
-        const debug = this.debugLogger.extend("updateGroupMembers");
-
-        let validMemberships: GraphMembership[] = [];
-
-        // Adding new memberships
-        if (members.length > 0) {
-
-            validMemberships = await this.graphHelper.addGroupMemberships(group, members);
-
-        }
-
-        const obsoleteMemberships: GraphMembership[] = await this.graphHelper.getObsoleteGroupMemberships(group, validMemberships);
-
-        // Removing obsolete memberships
-        if (obsoleteMemberships.length > 0) {
-
-            await this.graphHelper.removeGroupMemberships(group, obsoleteMemberships);
-
-        }
-
-    }
-
     public async updatePermissions(project: TeamProject, policy: IProjectPermission): Promise<void> {
 
         const debug = this.debugLogger.extend("updatePermissions");
@@ -178,7 +118,7 @@ export class ProjectUpdater implements IProjectUpdater {
 
                 this.logger.log(`Updating <${groupName}> group permissions`);
 
-                await this.updateGroupPermissions(project.name!, targetGroup, group.permissions);
+                await this.securityHelper.updateGroupPermissions(project.name!, targetGroup, group.permissions);
 
             }
 
@@ -187,7 +127,7 @@ export class ProjectUpdater implements IProjectUpdater {
 
                 this.logger.log(`Updating <${groupName}> group members`);
 
-                await this.updateGroupMembers(group.members, targetGroup);
+                await this.graphHelper.updateGroupMembers(group.members, targetGroup);
 
             }
